move workers and fractal functions to its corresponding modules
