+++ fract4d/c/fract4dmodule.cpp
@@ -34,6 +34,8 @@
 #include "fract4dc/sites.h"
 #include "fract4dc/images.h"
 #include "fract4dc/calcs.h"
+#include "fract4dc/workers.h"
+#include "fract4dc/functions.h"
 
 struct module_state
 {
@@ -42,13 +44,6 @@ struct module_state
 
 // really should be in module_state - one day
 void *cmap_module_handle = NULL;
-
-typedef enum
-{
-    DELTA_X,
-    DELTA_Y,
-    TOPLEFT
-} vec_type_t;
 
 
 static arena_t
@@ -323,250 +318,57 @@ image_get_fate(PyObject *self, PyObject *args)
 }
 
 
-
-static void
-fw_delete(IFractWorker *worker)
-{
-    delete worker;
-}
-
-struct ffHandle
-{
-    PyObject *pyhandle;
-    fractFunc *ff;
-};
-
-static void
-ff_delete(struct ffHandle *ffh)
-{
-#ifdef DEBUG_CREATION
-    fprintf(stderr, "%p : FF : DTOR\n", ffh);
-#endif
-    delete ffh->ff;
-    Py_DECREF(ffh->pyhandle);
-    delete ffh;
-}
-
-static IFractWorker *
-fw_fromcapsule(PyObject *capsule)
-{
-    IFractWorker *worker = (IFractWorker *)PyCapsule_GetPointer(capsule, OBTYPE_WORKER);
-    return worker;
-}
-
-static void
-pyfw_delete(PyObject *pyworker)
-{
-    IFractWorker *worker = fw_fromcapsule(pyworker);
-    fw_delete(worker);
-}
+/*
+* workers
+*/
 
 static PyObject *
 fw_create(PyObject *self, PyObject *args)
 {
-    int nThreads;
-    pf_obj *pfo;
-    ColorMap *cmap;
-    IImage *im;
-    IFractalSite *site;
-
-    PyObject *pypfo, *pycmap, *pyim, *pysite;
-
-    if (!PyArg_ParseTuple(args, "iOOOO",
-                          &nThreads,
-                          &pypfo,
-                          &pycmap,
-                          &pyim,
-                          &pysite))
-    {
-        return NULL;
-    }
-
-    cmap = colormaps::cmap_fromcapsule(pycmap);
-    pfo = (loaders::pf_fromcapsule(pypfo))->pfo;
-    im = images::image_fromcapsule(pyim);
-    site = sites::site_fromcapsule(pysite);
-    if (!cmap || !pfo || !im || !im->ok() || !site)
-    {
-        return NULL;
-    }
-
-    IFractWorker *worker = IFractWorker::create(nThreads, pfo, cmap, im, site);
-
-    if (!worker->ok())
-    {
-        PyErr_SetString(PyExc_ValueError, "Error creating worker");
-        delete worker;
-        return NULL;
-    }
-
-    PyObject *pyret = PyCapsule_New(worker, OBTYPE_WORKER, pyfw_delete);
-
-    return pyret;
+    return workers::fw_create(self, args);
 }
 
 static PyObject *
 fw_pixel(PyObject *self, PyObject *args)
 {
-    PyObject *pyworker;
-    int x, y, w, h;
-
-    if (!PyArg_ParseTuple(args, "Oiiii",
-                          &pyworker,
-                          &x, &y, &w, &h))
-    {
-        return NULL;
-    }
-
-    IFractWorker *worker = fw_fromcapsule(pyworker);
-    worker->pixel(x, y, w, h);
-
-    Py_INCREF(Py_None);
-    return Py_None;
+    return workers::fw_pixel(self, args);
 }
 
 static PyObject *
 fw_pixel_aa(PyObject *self, PyObject *args)
 {
-    PyObject *pyworker;
-    int x, y;
-
-    if (!PyArg_ParseTuple(args, "Oii",
-                          &pyworker,
-                          &x, &y))
-    {
-        return NULL;
-    }
-
-    IFractWorker *worker = fw_fromcapsule(pyworker);
-    worker->pixel_aa(x, y);
-
-    Py_INCREF(Py_None);
-    return Py_None;
+    return workers::fw_pixel_aa(self, args);
 }
 
 static PyObject *
 fw_find_root(PyObject *self, PyObject *args)
 {
-    PyObject *pyworker;
-    dvec4 eye, look;
-
-    if (!PyArg_ParseTuple(args, "O(dddd)(dddd)",
-                          &pyworker,
-                          &eye[VX], &eye[VY], &eye[VZ], &eye[VW],
-                          &look[VX], &look[VY], &look[VZ], &look[VW]))
-    {
-        return NULL;
-    }
-
-    IFractWorker *worker = fw_fromcapsule(pyworker);
-    dvec4 root;
-    int ok = worker->find_root(eye, look, root);
-
-    return Py_BuildValue(
-        "i(dddd)",
-        ok, root[0], root[1], root[2], root[3]);
+    return workers::fw_find_root(self, args);
 }
 
-static ffHandle *
-ff_fromcapsule(PyObject *pyff)
-{
-    ffHandle *ff = (ffHandle *)PyCapsule_GetPointer(pyff, OBTYPE_FFH);
-    if (NULL == ff)
-    {
-        fprintf(stderr, "%p : FF : CTOR\n", ff);
-    }
-    return ff;
-}
 
-static void
-pyff_delete(PyObject *pyff)
-{
-    ffHandle *ff = (ffHandle *)PyCapsule_GetPointer(pyff, OBTYPE_FFH);
-    ff_delete(ff);
-}
+/*
+* functions
+*/
 
 static PyObject *
 ff_create(PyObject *self, PyObject *args)
 {
-    PyObject *pypfo, *pycmap, *pyim, *pysite, *pyworker;
-    double params[N_PARAMS];
-    int eaa = -7, maxiter = -8, nThreads = -9;
-    int auto_deepen, periodicity;
-    int yflip;
-    render_type_t render_type;
-    pf_obj *pfo;
-    ColorMap *cmap;
-    IImage *im;
-    IFractalSite *site;
-    IFractWorker *worker;
-    int auto_tolerance;
-    double tolerance;
-
-    if (!PyArg_ParseTuple(
-            args,
-            "(ddddddddddd)iiiiOOiiiOOOid",
-            &params[0], &params[1], &params[2], &params[3],
-            &params[4], &params[5], &params[6], &params[7],
-            &params[8], &params[9], &params[10],
-            &eaa, &maxiter, &yflip, &nThreads,
-            &pypfo, &pycmap,
-            &auto_deepen,
-            &periodicity,
-            &render_type,
-            &pyim, &pysite,
-            &pyworker,
-            &auto_tolerance, &tolerance))
-    {
-        return NULL;
-    }
-
-    cmap = colormaps::cmap_fromcapsule(pycmap);
-    pfo = (loaders::pf_fromcapsule(pypfo))->pfo;
-    im = images::image_fromcapsule(pyim);
-    site = sites::site_fromcapsule(pysite);
-    worker = fw_fromcapsule(pyworker);
-
-    if (!cmap || !pfo || !im || !site || !worker)
-    {
-        return NULL;
-    }
-
-    fractFunc *ff = new fractFunc(
-        params,
-        eaa,
-        maxiter,
-        nThreads,
-        auto_deepen,
-        auto_tolerance,
-        tolerance,
-        yflip,
-        periodicity,
-        render_type,
-        -1, // warp_param
-        worker,
-        im,
-        site);
-
-    if (!ff)
-    {
-        return NULL;
-    }
-
-    ffHandle *ffh = new struct ffHandle;
-    ffh->ff = ff;
-    ffh->pyhandle = pyworker;
-
-#ifdef DEBUG_CREATION
-    fprintf(stderr, "%p : FF : CTOR\n", ffh);
-#endif
-
-    PyObject *pyret = PyCapsule_New(ffh, OBTYPE_FFH, pyff_delete);
-
-    Py_INCREF(pyworker);
-
-    return pyret;
+    return functions::ff_create(self, args);
 }
+
+static PyObject *
+ff_get_vector(PyObject *self, PyObject *args)
+{
+    return functions::ff_get_vector(self, args);
+}
+
+static PyObject *
+ff_look_vector(PyObject *self, PyObject *args)
+{
+    return functions::ff_look_vector(self, args);
+}
+
 
 static PyObject *
 rot_matrix(PyObject *self, PyObject *args)
@@ -613,88 +415,6 @@ eye_vector(PyObject *self, PyObject *args)
     return Py_BuildValue(
         "(dddd)",
         eyevec[0], eyevec[1], eyevec[2], eyevec[3]);
-}
-
-static PyObject *
-ff_get_vector(PyObject *self, PyObject *args)
-{
-    int vec_type;
-    PyObject *pyFF;
-
-    if (!PyArg_ParseTuple(
-            args,
-            "Oi",
-            &pyFF, &vec_type))
-    {
-        return NULL;
-    }
-
-    struct ffHandle *ffh = ff_fromcapsule(pyFF);
-    if (ffh == NULL)
-    {
-        return NULL;
-    }
-
-    fractFunc *ff = ffh->ff;
-    if (ff == NULL)
-    {
-        return NULL;
-    }
-
-    dvec4 vec;
-    switch (vec_type)
-    {
-    case DELTA_X:
-        vec = ff->deltax;
-        break;
-    case DELTA_Y:
-        vec = ff->deltay;
-        break;
-    case TOPLEFT:
-        vec = ff->topleft;
-        break;
-    default:
-        PyErr_SetString(PyExc_ValueError, "Unknown vector requested");
-        return NULL;
-    }
-
-    return Py_BuildValue(
-        "(dddd)",
-        vec[0], vec[1], vec[2], vec[3]);
-
-    return NULL;
-}
-
-static PyObject *
-ff_look_vector(PyObject *self, PyObject *args)
-{
-    PyObject *pyFF;
-    double x, y;
-    if (!PyArg_ParseTuple(
-            args,
-            "Odd",
-            &pyFF, &x, &y))
-    {
-        return NULL;
-    }
-
-    struct ffHandle *ffh = ff_fromcapsule(pyFF);
-    if (ffh == NULL)
-    {
-        return NULL;
-    }
-
-    fractFunc *ff = ffh->ff;
-    if (ff == NULL)
-    {
-        return NULL;
-    }
-
-    dvec4 lookvec = ff->vec_for_point(x, y);
-
-    return Py_BuildValue(
-        "(dddd)",
-        lookvec[0], lookvec[1], lookvec[2], lookvec[3]);
 }
 
 static PyObject *
