+++ setup.py
@@ -87,6 +87,8 @@ fract4d_sources = [
     'fract4d/c/fract4dc/sites.cpp',
     'fract4d/c/fract4dc/images.cpp',
     'fract4d/c/fract4dc/calcs.cpp',
+    'fract4d/c/fract4dc/workers.cpp',
+    'fract4d/c/fract4dc/functions.cpp',
 
     'fract4d/c/cmap.cpp',
     'fract4d/c/pointFunc.cpp',
